backend/requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
python-dotenv>=1.0.0
pydantic==2.5.3
livekit-agents[groq,deepgram,cartesia]>=1.0.0
livekit-plugins-deepgram==1.0.0
livekit-plugins-groq==1.0.0
livekit-plugins-cartesia==1.0.0
livekit-server-sdk-python==1.2.0
python-multipart==0.0.6 